@@ -1,27 +1,35 @@
-import Vue from "vue"
+import { createApp } from 'vue'
 import App from '@/App.vue'
 import router from '@/router'
 import store from '@/store/index'
 import ApiService from "@/services/ApiService"
-import 'bootstrap/dist/css/bootstrap.css'
-import 'bootstrap-vue/dist/bootstrap-vue.css'
-import { BootstrapVue, BootstrapVueIcons } from 'bootstrap-vue'
+import 'bootstrap/dist/css/bootstrap.min.css'
+import 'bootstrap-vue-next/dist/bootstrap-vue-next.css'
 
-Vue.config.productionTip = false;
-Vue.use(BootstrapVue)
-Vue.use(BootstrapVueIcons)
+import { BNavbar, BNavbarBrand, BNavbarNav, BNavItem, BCollapse, BModal, BButton, BContainer, BRow, modalManagerPlugin } from 'bootstrap-vue-next'
 
 ApiService.init();
 
-router.beforeEach((to, from, next) =>  {
-  Promise.all([store.dispatch("checkAuth")]).then(next)
-});
+// Restore Vuex auth state from localStorage
+const token = localStorage.getItem('token');
+const username = localStorage.getItem('username');
+if (token && username) {
+  store.commit('auth/setUser', { username, token });
+}
 
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-  data: {
-    ipAddress: 'localhost',
-  },
-}).$mount('#app');
+const app = createApp(App)
+app.use(router)
+app.use(store)
+app.use(modalManagerPlugin)
+
+app.component('BNavbar', BNavbar)
+app.component('BNavbarBrand', BNavbarBrand)
+app.component('BNavbarNav', BNavbarNav)
+app.component('BNavItem', BNavItem)
+app.component('BCollapse', BCollapse)
+app.component('BModal', BModal)
+app.component('BButton', BButton)
+app.component('BContainer', BContainer)
+app.component('BRow', BRow)
+
+app.mount('#app')
