@@ -1,4 +1,3 @@
-import datetime
 import os
 
 basedir = os.path.abspath(os.path.dirname(__file__))
@@ -8,7 +7,7 @@ class Config(object):
     SQLALCHEMY_ECHO = False
     SECRET_KEY = "UBERS3Cr3tKeyBr0"
     JWT_SECRET_KEY = "Kj31dsh4sF0uNDH1STE4SP0ON!"
-    SQLALCHEMY_DATABASE_URI = "sqlite:///{}/database/Application.db".format(basedir)
+    SQLALCHEMY_DATABASE_URI = f"sqlite:///{os.path.join(basedir, 'Application.db')}"
     FLASK_ADMIN_SWATCH = 'cerulean'
 
 class DevelopmentConfig(Config):
