@@ -1,36 +1,35 @@
-import axios from 'axios'
-import Vue from 'vue';
-import VueAxios from 'vue-axios'
-import JwtService from '@/services/JwtService'
+import axios from "axios";
 import { API_URL } from "@/common/config";
 
-class ApiService {
-	init() {
-		Vue.use(VueAxios, axios);
-		Vue.axios.defaults.baseURL = API_URL;
-	}
+const ApiService = {
+  init() {
+    // Set the base URL for axios
+    axios.defaults.baseURL = API_URL;
+    // Set token from localStorage if available
+    const token = localStorage.getItem('token');
+    if (token) {
+      this.setHeader(token);
+    }
+  },
+  setHeader(token) {
+    if (token) {
+      axios.defaults.headers.common["Authorization"] = `Bearer ${token}`;
+    } else {
+      delete axios.defaults.headers.common["Authorization"];
+    }
+  },
+  get(resource) {
+    return axios.get(resource);
+  },
+  post(resource, data) {
+    return axios.post(resource, data);
+  },
+  put(resource, data) {
+    return axios.put(resource, data);
+  },
+  delete(resource) {
+    return axios.delete(resource);
+  }
+};
 
-	setHeader() {
-		Vue.axios.defaults.headers.common[
-			"Authorization"
-		] = `Bearer ${JwtService.getToken()}`;
-	}
-
-	get(resource) {
-		return Vue.axios.get(`${resource}`);
-	}
-
-	post(resource, params) {
-		return Vue.axios.post(`${resource}`, params);
-	}
-
-	put(resource, params) {
-		return Vue.axios.put(`${resource}`, params);
-	}
-
-	delete(resource) {
-		return Vue.axios.delete(`${resource}`);
-	}
-}
-
-export default new ApiService();
+export default ApiService;
