@@ -1,24 +1,30 @@
 <template>
-  <b-row align-h="center">
-    <b-col cols="8">
-      <b-tabs content-class="mt-3">
-        <b-tab title="Login">
-          <LoginForm></LoginForm>
-        </b-tab>
-        <b-tab title="Register">
-          <RegisterForm></RegisterForm>
-        </b-tab>
-      </b-tabs>
-    </b-col>
-  </b-row>
+  <BRow align-h="center">
+    <BCol cols="8">
+      <BTabs content-class="mt-3">
+        <BTab title="Login">
+          <LoginForm />
+        </BTab>
+        <BTab title="Register">
+          <RegisterForm />
+        </BTab>
+      </BTabs>
+    </BCol>
+  </BRow>
 </template>
 
 <script>
+import { BRow, BCol, BTabs, BTab } from 'bootstrap-vue-next'
 import LoginForm from '@/components/user/LoginForm.vue'
 import RegisterForm from '@/components/user/RegisterForm.vue'
+
 export default {
   name: 'authComponent',
   components: {
+    BRow,
+    BCol,
+    BTabs,
+    BTab,
     LoginForm,
     RegisterForm
   },
